pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "myid-sdk"
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "PARAMETA Corp", email = "t_pds@parametacorp.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "icon-did-sdk @ git+https://git@github.com/iconloop/did-sdk-python.git@develop",
  "iconsdk~=2.4.5",
  "loguru==0.7.0",
  "requests==2.28.2",
  "pydantic[dotenv]==1.10.7",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "pytest~=7.2.2",
  "pytest-asyncio==0.21.1",
  "pytest-cov~=4.0.0",
  "pytest-mock~=3.10.0",
  "anyio[trio]~=3.7.0",
]
doc = [
  "mkdocs-material~=9.1.18",
]
all = [
  "myid-sdk[doc, dev]",
]

[project.urls]
Documentation = "https://github.com/iconloop/myid-sdk-python#readme"
Issues = "https://github.com/iconloop/myid-sdk-python/issues"
Source = "https://github.com/iconloop/myid-sdk-python"

[tool.hatch.version]
path = "myid/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=app {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["311"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "myid/__about__.py",
  "*/alembic/*",
  "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
